requirements.txt
faker==37.4.0
sqlalchemy==2.0.41
python-dotenv==1.1.0